+++ src/test/java/io/beanmapper/testmodel/collections/CollSourceConstruct.java
@@ -1,0 +1,14 @@
+package io.beanmapper.testmodel.collections;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import io.beanmapper.annotations.BeanCollection;
+import io.beanmapper.annotations.BeanCollectionUsage;
+
+public class CollSourceConstruct {
+
+    @BeanCollection(elementType = CollTarget.class, beanCollectionUsage = BeanCollectionUsage.CONSTRUCT)
+    public List<String> items = new ArrayList<>();
+
+}