+++ src/test/java/io/beanmapper/testmodel/collections/CollTarget.java
@@ -1,0 +1,9 @@
+package io.beanmapper.testmodel.collections;
+
+import java.util.List;
+
+public class CollTarget {
+
+    public List<String> items;
+
+}